Moved image routines to ImageTools
